+++ scim2_client/__init__.py
@@ -5,6 +5,7 @@ from .errors import ResponsePayloadValidationError
 from .errors import SCIMClientError
 from .errors import SCIMRequestError
 from .errors import SCIMResponseError
+from .errors import SCIMResponseErrorObject
 from .errors import UnexpectedContentFormat
 from .errors import UnexpectedContentType
 from .errors import UnexpectedStatusCode
@@ -14,6 +15,7 @@ __all__ = [
     "SCIMClientError",
     "SCIMRequestError",
     "SCIMResponseError",
+    "SCIMResponseErrorObject",
     "UnexpectedContentFormat",
     "UnexpectedContentType",
     "UnexpectedStatusCode",